blog-api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
 flask
flask_sqlalchemy
flask_migrate
flask_cors
psycopg2-binary
